moved test classes into their own files
