+++ tests/Deriver.Tests/Comparers/ClearanceDecisionCheckComparerTests.cs
@@ -1,6 +1,5 @@
 using Defra.TradeImportsDataApi.Domain.CustomsDeclaration;
 using Defra.TradeImportsDecisionDeriver.Deriver.Decisions.Comparers;
-using NSubstitute.ExceptionExtensions;
 
 namespace Defra.TradeImportsDecisionDeriver.Deriver.Tests.Comparers;
 
@@ -84,327 +83,5 @@ public class ClearanceDecisionCheckComparerTests
         var result = sut.Equals(check1, check2);
 
         result.Should().BeFalse();
-    }
-}
-
-public class ClearanceDecisionItemComparerTests
-{
-    [Fact]
-    public void HashCodeTest()
-    {
-        var item = new ClearanceDecisionItem
-        {
-            ItemNumber = 1,
-            Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-        };
-        var sut = new ClearanceDecisionItemComparer();
-
-        Action act = () => sut.GetHashCode(item);
-
-        act.Should().Throw<NotSupportedException>();
-    }
-
-    [Fact]
-    public void ReferenceEqualsTest()
-    {
-        var item = new ClearanceDecisionItem
-        {
-            ItemNumber = 1,
-            Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-        };
-        var sut = new ClearanceDecisionItemComparer();
-
-        var result = sut.Equals(item, item);
-
-        result.Should().BeTrue();
-    }
-
-    [Fact]
-    public void FirstItemIsNull()
-    {
-        var item = new ClearanceDecisionItem
-        {
-            ItemNumber = 1,
-            Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-        };
-        var sut = new ClearanceDecisionItemComparer();
-
-        var result = sut.Equals(null, item);
-
-        result.Should().BeFalse();
-    }
-
-    [Fact]
-    public void SecondItemIsNull()
-    {
-        var item = new ClearanceDecisionItem
-        {
-            ItemNumber = 1,
-            Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-        };
-        var sut = new ClearanceDecisionItemComparer();
-
-        var result = sut.Equals(item, null);
-
-        result.Should().BeFalse();
-    }
-
-    [Fact]
-    public void ReferenceDifferentButValuesSames()
-    {
-        var item1 = new ClearanceDecisionItem
-        {
-            ItemNumber = 1,
-            Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-        };
-        var item2 = new ClearanceDecisionItem
-        {
-            ItemNumber = 1,
-            Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-        };
-        var sut = new ClearanceDecisionItemComparer();
-
-        var result = sut.Equals(item1, item2);
-
-        result.Should().BeTrue();
-    }
-
-    [Fact]
-    public void ReferenceDifferentItemNumberDifferent()
-    {
-        var item1 = new ClearanceDecisionItem
-        {
-            ItemNumber = 1,
-            Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-        };
-        var item2 = new ClearanceDecisionItem
-        {
-            ItemNumber = 2,
-            Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-        };
-        var sut = new ClearanceDecisionItemComparer();
-
-        var result = sut.Equals(item1, item2);
-
-        result.Should().BeFalse();
-    }
-
-    [Fact]
-    public void ReferenceDifferentChecksDifferent()
-    {
-        var item1 = new ClearanceDecisionItem
-        {
-            ItemNumber = 1,
-            Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-        };
-        var item2 = new ClearanceDecisionItem
-        {
-            ItemNumber = 1,
-            Checks = [new ClearanceDecisionCheck { CheckCode = "H202", DecisionCode = "C03" }],
-        };
-        var sut = new ClearanceDecisionItemComparer();
-
-        var result = sut.Equals(item1, item2);
-
-        result.Should().BeFalse();
-    }
-
-    [Fact]
-    public void SameChecksDifferentOrder()
-    {
-        var item1 = new ClearanceDecisionItem
-        {
-            ItemNumber = 1,
-            Checks =
-            [
-                new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C04" },
-                new ClearanceDecisionCheck { CheckCode = "H202", DecisionCode = "C03" },
-            ],
-        };
-        var item2 = new ClearanceDecisionItem
-        {
-            ItemNumber = 1,
-            Checks =
-            [
-                new ClearanceDecisionCheck { CheckCode = "H202", DecisionCode = "C03" },
-                new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C04" },
-            ],
-        };
-        var sut = new ClearanceDecisionItemComparer();
-
-        var result = sut.Equals(item1, item2);
-
-        result.Should().BeTrue();
-    }
-}
-
-public class ClearanceDecisionComparerTests
-{
-    [Fact]
-    public void HashCodeTest()
-    {
-        var decision = new ClearanceDecision
-        {
-            SourceVersion = "1",
-            Items =
-            [
-                new ClearanceDecisionItem
-                {
-                    ItemNumber = 1,
-                    Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-                },
-            ],
-        };
-        var sut = new ClearanceDecisionComparer();
-
-        Action act = () => sut.GetHashCode(decision);
-
-        act.Should().Throw<NotSupportedException>();
-    }
-
-    [Fact]
-    public void ReferenceEqualsTest()
-    {
-        var decision = new ClearanceDecision
-        {
-            SourceVersion = "1",
-            Items =
-            [
-                new ClearanceDecisionItem
-                {
-                    ItemNumber = 1,
-                    Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-                },
-            ],
-        };
-        var sut = new ClearanceDecisionComparer();
-
-        var result = sut.Equals(decision, decision);
-
-        result.Should().BeTrue();
-    }
-
-    [Fact]
-    public void FirstItemIsNull()
-    {
-        var decision = new ClearanceDecision
-        {
-            SourceVersion = "1",
-            Items =
-            [
-                new ClearanceDecisionItem
-                {
-                    ItemNumber = 1,
-                    Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-                },
-            ],
-        };
-        var sut = new ClearanceDecisionComparer();
-
-        var result = sut.Equals(null, decision);
-
-        result.Should().BeFalse();
-    }
-
-    [Fact]
-    public void SecondItemIsNull()
-    {
-        var decision = new ClearanceDecision
-        {
-            SourceVersion = "1",
-            Items =
-            [
-                new ClearanceDecisionItem
-                {
-                    ItemNumber = 1,
-                    Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-                },
-            ],
-        };
-        var sut = new ClearanceDecisionComparer();
-
-        var result = sut.Equals(decision, null);
-
-        result.Should().BeFalse();
-    }
-
-    [Fact]
-    public void DifferentReferenceButSameValues()
-    {
-        var decision1 = new ClearanceDecision
-        {
-            SourceVersion = "1",
-            Items =
-            [
-                new ClearanceDecisionItem
-                {
-                    ItemNumber = 1,
-                    Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-                },
-            ],
-        };
-
-        var decision2 = new ClearanceDecision
-        {
-            SourceVersion = "1",
-            Items =
-            [
-                new ClearanceDecisionItem
-                {
-                    ItemNumber = 1,
-                    Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-                },
-            ],
-        };
-        var sut = new ClearanceDecisionComparer();
-
-        var result = sut.Equals(decision1, decision2);
-
-        result.Should().BeTrue();
-    }
-
-    [Fact]
-    public void DifferentReferenceButSameValuesDifferentOrder()
-    {
-        var decision1 = new ClearanceDecision
-        {
-            SourceVersion = "1",
-            Items =
-            [
-                new ClearanceDecisionItem
-                {
-                    ItemNumber = 1,
-                    Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-                },
-                new ClearanceDecisionItem
-                {
-                    ItemNumber = 2,
-                    Checks = [new ClearanceDecisionCheck { CheckCode = "H202", DecisionCode = "C04" }],
-                },
-            ],
-        };
-
-        var decision2 = new ClearanceDecision
-        {
-            SourceVersion = "1",
-            Items =
-            [
-                new ClearanceDecisionItem
-                {
-                    ItemNumber = 2,
-                    Checks = [new ClearanceDecisionCheck { CheckCode = "H202", DecisionCode = "C04" }],
-                },
-                new ClearanceDecisionItem
-                {
-                    ItemNumber = 1,
-                    Checks = [new ClearanceDecisionCheck { CheckCode = "H201", DecisionCode = "C03" }],
-                },
-            ],
-        };
-        var sut = new ClearanceDecisionComparer();
-
-        var result = sut.Equals(decision1, decision2);
-
-        result.Should().BeTrue();
     }
 }
